Remove the possibility to make Unique aliases
This commit modifies the Template type to add a lifetime parameter which
is used to prevent the creation of alias from a type that implements
Unique by giving the '_ lifetime to the paramter inside the new! macro.
This works because the lifetime can only be inferred when used in a
type parameter.

Doctests are added to demostrate this feature while also testing it.

The README is update to remove the warning about the aliasing problem